requirements.txt
aiohttp==1.3.0
appdirs==1.4.3
async-timeout==1.2.0
chardet==2.3.0
dj-database-url==0.4.2
Django==1.10.6
django-mptt==0.8.7
gunicorn==19.7.1
multidict==2.1.4
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
queuelib==1.4.2
six==1.10.0
telepot==10.5
urllib3==1.20
yarl==0.10.0
